[config] Add migration database connection
This additional connection allows to use a different SQL user to perform migrations.

As a result, this frees the main connection from these tasks, thus allowing the main SQL user to have only the permissions that are necessary for usual, day-to-day operations.

The migration user is optional. If its credentials are not defined, those of the main SQL user are used as a fallback.